+++ Assets/Scripts/Player.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: f1be3f663b4b4f3387fa110a2623b1b0
+timeCreated: 1652707875
