--- Assets/Scripts/PlayerSpawner.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using UnityEngine;
-
-public class PlayerSpawner : MonoBehaviour
-{
-    private void Awake()
-    {
-        
-    }
-}
