--- Assets/Scripts/PlayerSpawner.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 4bfd0108b83d4862a61245e6f50383c4
-timeCreated: 1652707038
