+++ Assets/Scripts/Terrain/Models/BlockType.cs
@@ -1,9 +1,20 @@
 namespace Terrain.Models
 {
-    public enum BlockType
+    public enum BlockType : byte
     {
+        /// <summary>
+        /// Represents void block
+        /// </summary>
         Nothing,
+        
+        /// <summary>
+        /// Represents air block
+        /// </summary>
         Air,
+        
+        /// <summary>
+        /// Represents grass block
+        /// </summary>
         Grass
     }
 }
